+++ just_audio/darwin/just_audio/Sources/just_audio/AudioPlayer.m
@@ -1538,10 +1538,18 @@ static dispatch_queue_t serialQueue = nil;
     }
 }
 
-- (void)dispose {
+- (void)dispose:(BOOL)calledFromDealloc {
     if (!_player) return;
     if (_processingState != none) {
         [_player pause];
+        [self updatePosition];
+        if (!calledFromDealloc) {
+            [self broadcastPlaybackEvent];
+        }
+        if (_playResult) {
+            _playResult(@{});
+            _playResult = nil;
+        }
         _processingState = none;
         // If used just before destroying the current FlutterEngine, this will result in:
         // NSInternalInconsistencyException: 'Sending a message before the FlutterEngine has been run.'